+++ rethinkdb-net/Interfaces/IChangefeedCompatibleQuery.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using System.ComponentModel;
-using RethinkDb.Spec;
 
 namespace RethinkDb
 {
     [ImmutableObject(true)]
-    public interface ITableQuery<T> : ISequenceQuery<T>
+    public interface IChangefeedCompatibleQuery<T> : IQuery
     {
     }
 }