+++ rethinkdb-net/Interfaces/IOrderByIndexQuery.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace RethinkDb
+{
+    public interface IOrderByIndexQuery<T> : ISequenceQuery<T>
+    {
+    }
+}